+++ packages/components/eslint.config.js
@@ -1,0 +1,3 @@
+import { createConfig } from '@boluo/eslint-config/base';
+
+export default createConfig(import.meta.dirname);
