+++ packages/components/package.json
@@ -1,31 +1,25 @@
 {
-  "name": "@boluo/interpreter",
+  "name": "@boluo/components",
   "version": "0.0.0",
   "dependencies": {
-    "@boluo/api": "0.0.0",
-    "@boluo/components": "0.0.0",
-    "@boluo/sort": "0.0.0",
-    "prando": "^6.0.1",
-    "react-intl": "^10.1.14"
+    "@boluo/types": "0.0.0"
   },
   "devDependencies": {
     "@boluo/eslint-config": "0.0.0",
     "@boluo/typescript-config": "0.0.0",
     "@types/node": "^26.1.1",
     "eslint": "^10.6.0",
-    "tsx": "^4.23.0",
     "typescript": "^6"
   },
   "exports": {
-    ".": "./src/index.ts"
+    "./*": "./src/*.ts"
   },
   "homepage": "https://github.com/mythal/boluo",
   "license": "AGPL-3.0",
   "private": true,
   "scripts": {
     "check": "tsc --noEmit",
-    "lint": "eslint .",
-    "test": "node --import tsx --test src/**/*.test.ts"
+    "lint": "eslint ."
   },
   "sideEffects": false,
   "type": "module"
