+++ packages/components/src/ref.ts
@@ -1,4 +1,4 @@
-import type { ComponentRef } from '@boluo/api';
+import type { ComponentRef } from '@boluo/types/bindings';
 
 export const componentRefsMatch = (requested: ComponentRef, actual: ComponentRef): boolean =>
   requested.scopeId === actual.scopeId &&
