+++ packages/ui/package.json
@@ -3,7 +3,7 @@
   "version": "0.0.0",
   "dependencies": {
     "@boluo/color": "0.0.0",
-    "@boluo/common": "0.0.0",
+    "@boluo/components": "0.0.0",
     "@boluo/icons": "0.0.0",
     "@boluo/types": "0.0.0",
     "@boluo/utils": "0.0.0",
